Revert "Add promoBanners widget (#19590)"
This reverts commit 455672901aec0b77d6bf615c061d1a55cc36c7db.
