+++ src/applications/debt-letters/tests/e2e/diary-codes-next-step.cypress.spec.js
@@ -2,7 +2,7 @@ import mockFeatureToggles from './fixtures/mocks/feature-toggles.json';
 import mockDebts from './fixtures/mocks/debts.json';
 import mockUser from './fixtures/mocks/mock-user.json';
 
-describe.skip('Diary Codes - Next Steps', () => {
+describe('Diary Codes - Next Steps', () => {
   beforeEach(() => {
     cy.login(mockUser);
     cy.intercept('GET', '/v0/feature_toggles*', mockFeatureToggles);
