+++ src/applications/medical-copays/tests/e2e/mcp.cypress.spec.js
@@ -2,7 +2,7 @@ import mockFeatureToggles from './fixtures/mocks/feature-toggles.json';
 import mockCopays from './fixtures/mocks/copays.json';
 import mockUser from './fixtures/mocks/mock-user.json';
 
-describe.skip('Medical Copays', () => {
+describe('Medical Copays', () => {
   const id = 'f4385298-08a6-42f8-a86f-50e97033fb85';
 
   beforeEach(() => {
