+++ src/platform/site-wide/banners/index.js
@@ -3,10 +3,26 @@ import React from 'react';
 import recordEvent from 'platform/monitoring/record-event';
 import Banner from '@department-of-veterans-affairs/component-library/Banner';
 // Relative imports.
-import MaintenanceBanner from './components/MaintenanceBanner';
 import startReactApp from '../../startup/react';
+import MaintenanceBanner from './components/MaintenanceBanner';
 import widgetTypes from '~/applications/static-pages/widgetTypes';
-import { deriveStorage } from './helpers';
+
+export const deriveStorage = banner => {
+  const dismissibleStatus = banner?.dataset?.dismissibleStatus;
+
+  // Don't have storage if the banner is not dismissible.
+  if (dismissibleStatus === 'perm') {
+    return undefined;
+  }
+
+  // If the banner is dismiss-session, we'll use sessionStorage.
+  if (dismissibleStatus === 'dismiss-session') {
+    return window.sessionStorage;
+  }
+
+  // Use localStorage by default.
+  return window.localStorage;
+};
 
 // Are you looking for where this is used?
 // Search for `data-widget-type="banner"` and `data-widget-type="maintenance-banner"` to find all the places this React widget is used.
