+++ src/platform/site-wide/banners/index.unit.spec.js
